# openapi/spec.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: Shoppy
  description: Common spec
servers:
  - url: /
    description: prod
paths:
...
  /goods/get_all:
    get:
      tags:
        - goods-controller
      summary: "Get paginated goods according filter query and pagination"
      operationId: getAllGoodsUsingFilterAndPagination
      x-spring-paginated: true
      parameters:
        - name: filter
          in: query
          required: false
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: size
          in: query
          required: false
          schema:
            type: integer
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommonGoodDto"
  /goods/get_by_id:
    post:
      tags:
        - goods-controller
      summary: "Get goods by ids list"
      operationId: getGoodsByIds
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: string
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommonGoodDto"
  /jwt/login:
    post:
      tags:
        - jwt-controller
      summary: "Do login"
      operationId: doLogin
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JwtRequestDto"
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JwtResponseDto"
        401:
          description: "Need authorization"
        500:
          description: "Error"
...
  /jwt/probe:
    get:
      tags:
        - jwt-controller
      summary: "Check, that authorization header contains valid token"
      operationId: probeToken
      responses:
        200:
          description: "OK"
          content:
            text/plain:
              schema:
                type: string
        500:
          description: "Error"
  /orders/create:
    post:
      tags:
        - orders-controller
      summary: "Place order"
      operationId: createOrder
      responses:
        201:
          description: "Order created, returns path to order view"
        401:
          description: "Credentials error"
        499:
          description: "Request error"
  /orders/get/{guid}:
    get:
      tags:
        - orders-controller
      summary: "Get order info"
      operationId: getOrderInfoByGuid
      parameters:
        - name: guid
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExtendedOrderDto"
        401:
          description: "Credentials error"
        499:
          description: "Request error"
...
  /user/roles/{login}:
    get:
      tags:
        - users-controller
      summary: "Get user roles"
      operationId: getUserRolesByLogin
      parameters:
        - name: login
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: "OK"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountRole"
components:
  schemas:
...
    JwtRequestDto:
      type: object
      properties:
        login:
          type: string
        password:
          type: string
    JwtResponseDto:
      type: object
      properties:
        token:
          type: string
        expirationTime:
          type: object
          x-type: java.time.OffsetDateTime
          x-defaultValue: java.time.OffsetDateTime.now()
    CommonGoodDto:
      type: object
      properties:
        name:
          type: string
        article:
          type: string
        status:
          $ref: "#/components/schemas/Status"
        price:
          type: number
    CommonOrderDto:
      type: object
      properties:
        id:
          type: integer
        info:
          type: string
        status:
          $ref: "#/components/schemas/Status"
    ExtendedOrderDto:
      allOf:
        - $ref: "#/components/schemas/CommonOrderDto"
        - type: object
          properties:
            guid:
              type: string
            entries:
              type: array
              items:
                $ref: "#/components/schemas/OrderEntry"
    OrderEntry:
      type: object
      properties:
        good:
          $ref: "#/components/schemas/CommonGoodDto"
        count:
          type: integer
    AccountRole:
      type: string
      enum:
        - "ROLE_ADMIN"
        - "ROLE_USER"
        - "ROLE_GUEST"
    Status:
      type: string
      enum:
        - "ADDED"
        - "ACTIVE"
        - "REMOVED"
        - "DISABLED"